+++ src/IClassParser.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Butterfly\Component\Annotations;
+
+/**
+ * @author Marat Fakhertdinov <marat.fakhertdinov@gmail.com>
+ */
+interface IClassParser
+{
+    /**
+     * @param string $className
+     * @return array
+     */
+    public function parseClass($className);
+}