+++ tests/unit/ClassFinder/TestDir/ClassFoo.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Butterfly\Component\Annotations\Tests\ClassFinder\TestDir;
+
+class ClassFoo
+{
+
+}