+++ tests/unit/FileLoader/TestDir/InnerTestDir/ClassBaz.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Butterfly\Component\Annotations\Tests\FileLoader\TestDir\InnerTestDir;
-
-class ClassBaz
-{
-
-}